// src/app/auth/_components/register/types.ts
export type RegisterFormData = {
  email: string;
  password: string;
  confirmPassword: string;
  firstName: string;
  middleName: string;
  lastName: string;
  nationalId: string;
  dateOfBirth: string;
  address: string;
  phoneNumber: string;
  studentId: string;
  disabilityStatus: string;
  revolutionaryContribution: string;
};